refactor: unify version verification into single script
- Consolidate verify-versions-sync.mts and verify-version-bump.mts into verify-version.mts
- Include root package.json in version sync check
- Update auto-bump-version workflow to also bump root package.json
- Sync all package.json versions to 0.0.0-alpha.12
- Update pre-commit config to use single verify-version hook
- Update CI to use unified script
- Update docs/practices/git.md
