+++ .github/workflows/auto-bump-version.yml
@@ -29,6 +29,9 @@ jobs:
           NEW_VERSION=$(node -e "console.log(require('./package.json').version)")
           echo "NEW_VERSION=$NEW_VERSION" >> $GITHUB_ENV
 
+      - name: Update root package version
+        run: npm version ${{ env.NEW_VERSION }} --no-git-tag-version
+
       - name: Update react package version and dependency
         run: |
           cd react
@@ -44,6 +47,6 @@ jobs:
         run: |
           git config user.name "github-actions[bot]"
           git config user.email "github-actions[bot]@users.noreply.github.com"
-          git add base/package.json react/package.json react/example/package.json
+          git add package.json base/package.json react/package.json react/example/package.json
           git commit -m "chore: bump version to ${{ env.NEW_VERSION }} [skip ci]"
           git push
