+++ .pre-commit-config.yaml
@@ -11,15 +11,9 @@ repos:
       - id: mixed-line-ending
   - repo: local
     hooks:
-      - id: verify-versions-sync
-        name: verify-versions-sync
-        entry: npx tsx scripts/verify-versions-sync.mts
-        language: system
-        pass_filenames: false
-        always_run: true
-      - id: verify-version-bump
-        name: verify-version-bump
-        entry: npx tsx scripts/verify-version-bump.mts
+      - id: verify-version
+        name: verify-version
+        entry: npx tsx scripts/verify-version.mts
         language: system
         pass_filenames: false
         always_run: true
