+++ react/example/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "html-flip-book-react-example",
-	"version": "0.0.0-alpha.10",
+	"version": "0.0.0-alpha.12",
 	"description": "Flipbook component for HTML",
 	"homepage": "https://doradsoft.github.io/html-flip-book/",
 	"type": "module",
