+++ react/package.json
@@ -1,7 +1,7 @@
 {
 	"name": "html-flip-book-react",
 	"description": "Flip Book React Component",
-	"version": "0.0.0-alpha.10",
+	"version": "0.0.0-alpha.12",
 	"type": "module",
 	"author": "DoradSoft",
 	"main": "./dist/flip-book.js",
